add sidebar to main.jsx, use context for getting username
