+++ frontend/src/context/UserContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+export const UserContext = createContext({
+    username: "User"
+});
